+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.0.3]
+* fixed non-final values in AccordionSection
+* changed default `scrollIntoView` animation to `fast`
+* updated docs
+
+## [1.0.2]
+* fixed Flutter SDK version
+
 ## [1.0.1]
 * fixed dependencies
 * added null safety